Take full ownership of this fork
Change the groupId to `<groupId>com.zartc.maven.plugin</groupId>`
Change the main packages name to `com.zartc.maven.plugin.gitflow`
