+++ src/main/java/com/zartc/maven/plugin/gitflow/GitFlowReleaseStartMojo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.amashchenko.maven.plugin.gitflow;
+package com.zartc.maven.plugin.gitflow;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -24,18 +24,18 @@ import org.codehaus.plexus.components.interactivity.PrompterException;
 import org.codehaus.plexus.util.StringUtils;
 import org.codehaus.plexus.util.cli.CommandLineException;
 
-import com.amashchenko.maven.plugin.gitflow.i18n.CommitMessages;
-import com.amashchenko.maven.plugin.gitflow.i18n.ErrorMessages;
-import com.amashchenko.maven.plugin.gitflow.i18n.PromptMessages;
+import com.zartc.maven.plugin.gitflow.i18n.CommitMessages;
+import com.zartc.maven.plugin.gitflow.i18n.ErrorMessages;
+import com.zartc.maven.plugin.gitflow.i18n.PromptMessages;
 
 /**
- * The git flow hotfix start mojo.
+ * The git flow release start mojo.
  *
  * @author Aleksandr Mashchenko
  *
  */
-@Mojo(name = "hotfix-start", aggregator = true)
-public class GitFlowHotfixStartMojo extends AbstractGitFlowMojo {
+@Mojo(name = "release-start", aggregator = true)
+public class GitFlowReleaseStartMojo extends AbstractGitFlowMojo {
 
     /** {@inheritDoc} */
     @Override
@@ -47,21 +47,28 @@ public class GitFlowHotfixStartMojo extends AbstractGitFlowMojo {
             // check uncommitted changes
             checkUncommittedChanges();
 
-            // need to be in master to get correct project version
-            // git checkout master
-            gitCheckout(gitFlowConfig.getProductionBranch());
+            // git for-each-ref --count=1 refs/heads/release/*
+            String pattern = "refs/heads/" + gitFlowConfig.getReleaseBranchPrefix() + "*";
+			final String releaseBranch = executeGitCommandReturn(
+                    "for-each-ref", "--count=1", pattern);
 
-            String defaultVersion = "1.0.1";
+            if (StringUtils.isNotBlank(releaseBranch)) {
+                throw new MojoFailureException(msg.getMessage(ErrorMessages.release_branch_already_exists));
+            }
+
+            // need to be in develop to get correct project version
+            // git checkout develop
+            gitCheckout(gitFlowConfig.getDevelopmentBranch());
 
             // get current project version from pom
             final String currentVersion = getCurrentProjectVersion();
 
-            // get default hotfix version
+            String defaultVersion = "1.0.0";
+            // get default release version
             try {
                 final DefaultVersionInfo versionInfo = new DefaultVersionInfo(
                         currentVersion);
-                defaultVersion = versionInfo.getNextVersion()
-                        .getReleaseVersionString();
+                defaultVersion = versionInfo.getReleaseVersionString();
             } catch (VersionParseException e) {
                 if (getLog().isDebugEnabled()) {
                     getLog().debug(e);
@@ -69,34 +76,28 @@ public class GitFlowHotfixStartMojo extends AbstractGitFlowMojo {
             }
 
             String version = null;
-            try {
-                version = prompter.prompt(
-                		msg.getMessage(PromptMessages.hotfix_branch_name_to_create_prompt, defaultVersion));
-            } catch (PrompterException e) {
-                getLog().error(e);
+            if (settings.isInteractiveMode()) {
+                try {
+                    version = prompter.prompt(
+                    		msg.getMessage(PromptMessages.release_branch_name_to_create_prompt, defaultVersion));
+                } catch (PrompterException e) {
+                    getLog().error(e);
+                }
             }
 
             if (StringUtils.isBlank(version)) {
                 version = defaultVersion;
             }
 
-            // git for-each-ref refs/heads/hotfix/...
-            String pattern = "refs/heads/" + gitFlowConfig.getHotfixBranchPrefix() + version;
-			final String hotfixBranch = executeGitCommandReturn("for-each-ref", pattern);
-
-            if (StringUtils.isNotBlank(hotfixBranch)) {
-                throw new MojoFailureException(msg.getMessage(ErrorMessages.hotfix_branch_name_duplicate));
-            }
-
-            // git checkout -b hotfix/... master
-            gitCreateAndCheckout(gitFlowConfig.getHotfixBranchPrefix()
-                    + version, gitFlowConfig.getProductionBranch());
+            // git checkout -b release/... develop
+            gitCreateAndCheckout(gitFlowConfig.getReleaseBranchPrefix() + version,
+            		gitFlowConfig.getDevelopmentBranch());
 
             // mvn versions:set -DnewVersion=... -DgenerateBackupPoms=false
             mvnSetVersions(version);
 
-            // git commit -a -m updating poms for hotfix
-            gitCommit(msg.getMessage(CommitMessages.updating_pom_for_hotfix_version, version));
+            // git commit -a -m updating poms for release
+            gitCommit(msg.getMessage(CommitMessages.updating_pom_for_release_version, version));
 
             if (installProject) {
                 // mvn clean install
