+++ src/main/java/com/zartc/maven/plugin/gitflow/i18n/ErrorMessages.java
@@ -1,11 +1,11 @@
-package com.amashchenko.maven.plugin.gitflow.i18n;
+package com.zartc.maven.plugin.gitflow.i18n;
 
 import ch.qos.cal10n.BaseName;
 import ch.qos.cal10n.Locale;
 import ch.qos.cal10n.LocaleData;
 
 
-@BaseName("com/amashchenko/maven/plugin/gitflow/i18n/ErrorMessages")
+@BaseName("com/zartc/maven/plugin/gitflow/i18n/ErrorMessages")
 @LocaleData({ @Locale("en") })
 public enum ErrorMessages {
 	unexpected_error,
