+++ src/main/java/com/zartc/maven/plugin/gitflow/i18n/PromptMessages.java
@@ -1,11 +1,11 @@
-package com.amashchenko.maven.plugin.gitflow.i18n;
+package com.zartc.maven.plugin.gitflow.i18n;
 
 import ch.qos.cal10n.BaseName;
 import ch.qos.cal10n.Locale;
 import ch.qos.cal10n.LocaleData;
 
 
-@BaseName("com/amashchenko/maven/plugin/gitflow/i18n/PromptMessages")
+@BaseName("com/zartc/maven/plugin/gitflow/i18n/PromptMessages")
 @LocaleData({ @Locale("en") })
 public enum PromptMessages {
 	feature_branch_list_header,
